SR_surface/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sr-surface"
version = "1.0.0"
description = "Surface Super-Resolution using 2D Swin Transformer"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache License 2.0"}
authors = [
    {name = "Surface SR Team", email = "sr-surface@example.com"}
]
keywords = ["super-resolution", "surface", "swin-transformer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "mindspore>=2.0.0",
    "numpy>=1.21.0",
    "pyyaml>=5.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "black>=21.7b0",
    "flake8>=3.9.0",
]
